[BE] refactor: DTO 잘못된 import 경로 수정 및 관련 로직 업데이트 #745
- 테스트 코드에서의 import 경로 수정.
