+++ backend/admin/src/main/java/com/onair/hearit/admin/dto/response/AdminRecommendHearitResponse.java
@@ -1,4 +1,4 @@
-package com.onair.hearit.admin.dto.request;
+package com.onair.hearit.admin.dto.response;
 
 import com.onair.hearit.core.domain.Hearit;
 import java.time.LocalDate;
